+++ tests/test_model.py
@@ -13,7 +13,35 @@ from .adapters import (
     run_scaled_dot_product_attention,
     run_transformer_block,
     run_transformer_lm,
+    run_linear, 
+    run_embedding,
 )
+
+
+def test_linear(numpy_snapshot, ts_state_dict, in_embeddings, d_model, d_ff):
+    w1_weight = ts_state_dict[0][f"layers.0.ffn.w1.weight"]
+    output = run_linear(
+        d_in=d_model,
+        d_out=d_ff,
+        weights=w1_weight,
+        in_features=in_embeddings,
+    )
+    numpy_snapshot.assert_match(
+        output
+    )
+
+
+def test_embedding(numpy_snapshot, ts_state_dict, in_indices, vocab_size, d_model):
+    embedding_weight = ts_state_dict[0][f"token_embeddings.weight"]
+    output = run_embedding(
+        vocab_size=vocab_size,
+        d_model=d_model,
+        weights=embedding_weight,
+        token_ids=in_indices,
+    )
+    numpy_snapshot.assert_match(
+        output
+    )
 
 
 def test_swiglu(numpy_snapshot, ts_state_dict, in_embeddings, d_model, d_ff):